Multiple channels activation and channel selection
The code allows the user to freely select a channel to perform a conversion, measuring the value of a placed resistor. For more information related to the architecture and connections, refer to the application note CN0383, multiple two-wire configuration. Consider that the code runs in the Raspbery Pi Pico board and uses the MycroPython tool.