SPI communication for ESP32 and ADC7124-8
Code for reading and writing the ADC7124-8 registers using the ESP32 microcontroller. The file is written to use MicroPython tool, that means that the main.py file have to be uploaded to the board using the Rshell tool, the results can be seen in the Arduino’s serial terminal at 115200 baud rate.